refac: Move JokeResponse and SearchResponse interfaces to new interfaces.tsx file in src directory
