+++ src/components/Header/Header.tsx
@@ -1,6 +1,6 @@
-import './Header.css'
 import { Link } from 'react-router-dom';
 import logo from '../../header-2.png';
+import './Header.css'
 
 const Header = () => {
   return (
